Replace Judy JSL trees with djb's critbit trees.
This gives speed-up for large zone (4 Mrecords) validation with small
number of signatures in the range of 15% (on a Core i7 CPU) to 36% (on a
cheap AMD CPU).

Judy is still used for JL (storing RR sets by type), and for JHS (NSEC3
hashes).  The latter might be replaced with critbit in the near future
as well, the former - with something else, eventually.

I used my own adaptation of public domain critbit code from
https://github.com/agl/critbit.  The changes are:

- de-cweb-ification
- support for user data storage
- cbtree_next() implemented
- standard memory allocation replaced with a faster
  "never frees a bit" version